Added my name to the README.md
